+++ src/Game.h
@@ -56,7 +56,7 @@ struct TitleState {
 
 struct HighScoreState {
 	const char* scoreType[Constants::MaxHighScores_];
-	float scores[Constants::MaxHighScores_];
+	int scores[Constants::MaxHighScores_];
 	_TTF_Font* font;
 };
